Switched widget to JobIntentService
