+++ app/src/main/AndroidManifest.xml
@@ -118,7 +118,8 @@
         </activity>
 
         <service
-            android:name=".service.PlayerWidgetService"
+            android:name=".service.PlayerWidgetJobService"
+            android:permission="android.permission.BIND_JOB_SERVICE"
             android:enabled="true"
             android:exported="false">
         </service>
