+++ app/src/main/java/de/danoeh/antennapod/receiver/PlayerWidget.java
@@ -5,14 +5,13 @@ import android.appwidget.AppWidgetProvider;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.support.v4.content.ContextCompat;
 import android.text.TextUtils;
 import android.util.Log;
 
 import java.util.Arrays;
 
 import de.danoeh.antennapod.core.service.playback.PlaybackService;
-import de.danoeh.antennapod.service.PlayerWidgetService;
+import de.danoeh.antennapod.service.PlayerWidgetJobService;
 
 public class PlayerWidget extends AppWidgetProvider {
     private static final String TAG = "PlayerWidget";
@@ -23,17 +22,7 @@ public class PlayerWidget extends AppWidgetProvider {
     public void onReceive(Context context, Intent intent) {
         Log.d(TAG, "onReceive");
         super.onReceive(context, intent);
-        // don't do anything if we're not enabled
-        if (!isEnabled(context)) {
-            return;
-        }
-
-        // these come from the PlaybackService when things should get updated
-        if (TextUtils.equals(intent.getAction(), PlaybackService.FORCE_WIDGET_UPDATE)) {
-            startUpdate(context);
-        } else if (TextUtils.equals(intent.getAction(), PlaybackService.STOP_WIDGET_UPDATE)) {
-            stopUpdate(context);
-        }
+        PlayerWidgetJobService.updateWidget(context);
     }
 
     @Override
@@ -41,14 +30,13 @@ public class PlayerWidget extends AppWidgetProvider {
         super.onEnabled(context);
         Log.d(TAG, "Widget enabled");
         setEnabled(context, true);
-        startUpdate(context);
+        PlayerWidgetJobService.updateWidget(context);
     }
 
     @Override
-    public void onUpdate(Context context, AppWidgetManager appWidgetManager,
-                         int[] appWidgetIds) {
+    public void onUpdate(Context context, AppWidgetManager appWidgetManager, int[] appWidgetIds) {
         Log.d(TAG, "onUpdate() called with: " + "context = [" + context + "], appWidgetManager = [" + appWidgetManager + "], appWidgetIds = [" + Arrays.toString(appWidgetIds) + "]");
-        startUpdate(context);
+        PlayerWidgetJobService.updateWidget(context);
     }
 
     @Override
@@ -56,20 +44,9 @@ public class PlayerWidget extends AppWidgetProvider {
         super.onDisabled(context);
         Log.d(TAG, "Widget disabled");
         setEnabled(context, false);
-        stopUpdate(context);
     }
 
-    private void startUpdate(Context context) {
-        Log.d(TAG, "startUpdate() called with: " + "context = [" + context + "]");
-        ContextCompat.startForegroundService(context, new Intent(context, PlayerWidgetService.class));
-    }
-
-    private void stopUpdate(Context context) {
-        Log.d(TAG, "stopUpdate() called with: " + "context = [" + context + "]");
-        context.stopService(new Intent(context, PlayerWidgetService.class));
-    }
-
-    private boolean isEnabled(Context context) {
+    public static boolean isEnabled(Context context) {
         SharedPreferences prefs = context.getSharedPreferences(PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getBoolean(KEY_ENABLED, false);
     }
